Charts: AreaChart reuses LineChart's renderDefaultTooltip
Drops the duplicate default-tooltip + AreaChartTooltipDatum copies. The
shape is identical to LineChart's, so AreaChart now imports from there.
Also trims a few overly-prose code comments that were saying twice what
the variable name already says, and removes the unused tooltip CSS rules
that came along with the deleted file.

- line-chart.tsx: export `renderDefaultTooltip` (was module-private).
- line-chart/index.ts: re-export it.
- area-chart.tsx: import from '../line-chart' instead of './private'.
- area-chart/private/default-tooltip.tsx: deleted.
- area-chart/private/validate-data.ts: split out so private/index.ts can
  still surface it.
- area-chart/types.ts + index.ts + src/index.ts: drop AreaChartTooltipDatum.
- area-chart.module.scss: drop now-unused __tooltip* rules.
