+++ projects/js-packages/charts/src/charts/area-chart/area-chart.module.scss
@@ -13,27 +13,6 @@
 	svg {
 		overflow: visible;
 	}
-
-	// Background is set by visx's TooltipInPortal from the active theme
-	// (see <Tooltip> in @visx/xychart). Don't override it here — that would
-	// break ThemeProvider-driven theming (e.g. dark mode).
-	&__tooltip {
-		padding: var(--wpds-dimension-padding-sm, 8px);
-	}
-
-	&__tooltip-date {
-		font-weight: var(--wpds-typography-font-weight-medium, 499);
-		padding-bottom: var(--wpds-dimension-padding-md, 12px);
-	}
-
-	&__tooltip-row {
-		padding: var(--wpds-dimension-padding-xs, 4px) 0;
-	}
-
-	&__tooltip-label {
-		font-weight: var(--wpds-typography-font-weight-medium, 499);
-		padding-right: var(--wpds-dimension-padding-lg, 16px);
-	}
 }
 
 @keyframes rise {
